Eliminate small_vector_adaptor & small_vector_facade Implement agency::point with array + arithmetic_tuple_facade
Fixes #317
